More refactoring into separate tests If all functions have their own individual test files, it should be a lot easier to identify which functions aren't covered by tests yet
