+++ tests/PhpSpreadsheetTests/Calculation/Functions/Statistical/AverageTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Statistical;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Statistical;
+use PHPUnit\Framework\TestCase;
+
+class AverageTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerAVERAGE
+     *
+     * @param mixed $expectedResult
+     */
+    public function testAVERAGE($expectedResult, ...$args)
+    {
+        $result = Statistical::AVERAGE(...$args);
+        self::assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerAVERAGE()
+    {
+        return require 'data/Calculation/Statistical/AVERAGE.php';
+    }
+}
