+++ tests/data/Calculation/Statistical/CRITBINOM.php
@@ -1,0 +1,24 @@
+<?php
+
+return [
+    [
+        4,
+        6, 0.5, 0.75,
+    ],
+    [
+        7,
+        12, 0.5, 0.75,
+    ],
+    [
+        46,
+        100, 0.5, 0.2,
+    ],
+    [
+        50,
+        100, 0.5, 0.5,
+    ],
+    [
+        56,
+        100, 0.5, 0.9,
+    ],
+];
